+++ packages/angular-workspace/nimble-angular/table-column/mapping/tests/nimble-table-column-mapping.directive.spec.ts
@@ -34,6 +34,7 @@ describe('NimbleTableColumnMapping', () => {
                         min-pixel-width="40"
                         sort-direction="${TableColumnSortDirection.ascending}"
                         sort-index="0"
+                        sorting-disabled
                         group-index="0"
                         grouping-disabled
                         width-mode="icon-size"
@@ -101,6 +102,11 @@ describe('NimbleTableColumnMapping', () => {
             expect(nativeElement.sortIndex).toBe(0);
         });
 
+        it('will use template string value for sortingDisabled', () => {
+            expect(directive.sortingDisabled).toBeTrue();
+            expect(nativeElement.sortingDisabled).toBeTrue();
+        });
+
         it('will use template string values for fractionalWidth', () => {
             expect(directive.fractionalWidth).toBe(2);
             expect(nativeElement.fractionalWidth).toBe(2);
@@ -143,6 +149,7 @@ describe('NimbleTableColumnMapping', () => {
                         [min-pixel-width]="minPixelWidth"
                         [sort-direction]="sortDirection"
                         [sort-index]="sortIndex"
+                        [sorting-disabled]="sortingDisabled"
                         [group-index]="groupIndex"
                         [grouping-disabled]="groupingDisabled"
                         [width-mode]="widthMode"
@@ -163,6 +170,7 @@ describe('NimbleTableColumnMapping', () => {
             public columnHidden = true;
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
             public sortIndex: number | null = 0;
+            public sortingDisabled = false;
             public groupIndex: number | null = 0;
             public groupingDisabled = false;
             public widthMode: TableColumnMappingWidthMode = TableColumnMappingWidthMode.iconSize;
@@ -282,6 +290,17 @@ describe('NimbleTableColumnMapping', () => {
             expect(nativeElement.sortIndex).toBe(null);
         });
 
+        it('can be configured with property binding for sortingDisabled', () => {
+            expect(directive.sortingDisabled).toBeFalse();
+            expect(nativeElement.sortingDisabled).toBeFalse();
+
+            fixture.componentInstance.sortingDisabled = true;
+            fixture.detectChanges();
+
+            expect(directive.sortingDisabled).toBeTrue();
+            expect(nativeElement.sortingDisabled).toBeTrue();
+        });
+
         it('can be configured with property binding for fractionalWidth', () => {
             expect(directive.fractionalWidth).toBe(2);
             expect(nativeElement.fractionalWidth).toBe(2);
@@ -387,6 +406,7 @@ describe('NimbleTableColumnMapping', () => {
                         [attr.min-pixel-width]="minPixelWidth"
                         [attr.sort-direction]="sortDirection"
                         [attr.sort-index]="sortIndex"
+                        [attr.sorting-disabled]="sortingDisabled"
                         [attr.group-index]="groupIndex"
                         [attr.grouping-disabled]="groupingDisabled"
                         [attr.width-mode]="widthMode"
@@ -407,6 +427,7 @@ describe('NimbleTableColumnMapping', () => {
             public columnHidden = true;
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
             public sortIndex: number | null = 0;
+            public sortingDisabled = false;
             public groupIndex: number | null = 0;
             public groupingDisabled = false;
             public widthMode: TableColumnMappingWidthMode = TableColumnMappingWidthMode.iconSize;
@@ -524,6 +545,17 @@ describe('NimbleTableColumnMapping', () => {
 
             expect(directive.sortIndex).toBe(null);
             expect(nativeElement.sortIndex).toBe(null);
+        });
+
+        it('can be configured with attribute binding for sortingDisabled', () => {
+            expect(directive.sortingDisabled).toBeFalse();
+            expect(nativeElement.sortingDisabled).toBeFalse();
+
+            fixture.componentInstance.sortingDisabled = true;
+            fixture.detectChanges();
+
+            expect(directive.sortingDisabled).toBeTrue();
+            expect(nativeElement.sortingDisabled).toBeTrue();
         });
 
         it('can be configured with attribute binding for fractionalWidth', () => {