+++ packages/nimble-components/src/table/index.ts
@@ -515,7 +515,7 @@ export class Table<
         column: TableColumn,
         allowMultiSort: boolean
     ): void {
-        if (column.sortingDisabled) {
+        if (column.columnInternals.sortingDisabled) {
             return;
         }
 
@@ -804,7 +804,7 @@ export class Table<
 
     private getColumnsParticipatingInSorting(): TableColumn[] {
         return this.columns.filter(
-            x => !x.sortingDisabled
+            x => !x.columnInternals.sortingDisabled
                 && x.columnInternals.currentSortDirection
                     !== TableColumnSortDirection.none
                 && typeof x.columnInternals.currentSortIndex === 'number'