+++ packages/nimble-components/src/table/models/table-update-tracker.ts
@@ -135,6 +135,7 @@ export class TableUpdateTracker<
         } else if (
             isColumnInternalsProperty(
                 changedColumnProperty,
+                'sortingDisabled',
                 'currentSortDirection',
                 'currentSortIndex'
             )